* master:
  [new_cadofactor.pl]
  [lingen_binary.cpp] added cast to avoid warning with gcc 4.0.2
  [new_cadofactor.pl]
  again and again
  groumf!
  better at last?
  missing ?
  minor modifs.
  [merge.c, merge_mono.c] my_cost now returns a double, which avoids overflow
  More cleaning.
  More cleaning and reorganization. USE_TAB == 0 is being removed everywhere.
  Started necessary cleaning of swar vs. merge stuff, preparing for
  preparing next move.
  A header for balance.c, and a description of usage.
  Have param_list library accept knobs of the form "-verbose" or

+++ linalg/merge.c
@@ -16,6 +16,9 @@
 #include "files.h"
 #include "gzip.h"
 #include "sparse.h"
+#include "dclist.h"
+#include "sparse_mat.h"
+#include "swar.h"
 #include "merge_mono.h"
 #include "prune.h"
 
@@ -199,10 +202,9 @@ main(int argc, char *argv[])
 #endif
 
     gzip_close(rep.outfile, outname);
-    fprintf(stderr, "Final matrix has N=%d nc=%d (%d) w(M)=%lu N*w(M)=%lu\n",
+    fprintf(stderr, "Final matrix has N=%d nc=%d (%d) w(M)=%lu N*w(M)=%1.0f\n",
 	    mat.rem_nrows, mat.rem_ncols, mat.rem_nrows-mat.rem_ncols,
-	    mat.weight,
-	    ((unsigned long)mat.rem_nrows) * mat.weight);
+	    mat.weight, (double) mat.rem_nrows * (double) mat.weight);
 #if TEX
     fprintf(stderr, "\\end{verbatim}\n");
 #endif
